// api/src/instagram/channel/dto/channel-reel.dto.ts
export default interface ChannelReelDTO {
    code?: string;
    urL?: string
    channel_reel_numerical_order?: number;
    comment_count?: number;
    id?: string;
    image_height?: number;
    image_width?: number;
    image_url?: string;
    like_count?: number;
    media_type?: number;
    pk?: string;
    play_count?: number;
    product_type?: string;
    video_url?: string;
    channel_username?: string;
}